requirements.txt
python-dotenv~=0.20.0
discord.py
discord-py-slash-command
